Moved WaypointNavigationActivity from the SDK to the test app
