+++ app/src/main/AndroidManifest.xml
@@ -51,7 +51,7 @@
         </activity>
 
         <activity
-            android:name="com.mapbox.services.android.navigation.ui.v5.WaypointNavigationActivity"
+            android:name=".activity.WaypointNavigationActivity"
             android:label="@string/title_waypoint_navigation">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
